Add UpdateController

+++ Models/OrderDetail.cs
@@ -16,9 +16,11 @@ namespace webapp.Models{
 
         public string ItemID { get; set; }
 
-        public float Quantity { get; set; }
+        // public float Quantity { get; set; }
+        public double Quantity { get; set; }
 
-        public float Price { get; set; }
+        // public float Price { get; set; }
+        public double Price { get; set; }
         
         public decimal Amount { get; set; }
